+++ database/migrations/2023_10_03_183828_create_reservas_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('reservas', function (Blueprint $table) {
             $table->id();
-            $table->integer('mesa_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('mesa_id');
+            $table->unsignedBigInteger('user_id');
             $table->date('fecha');
             $table->time('hora');
             $table->boolean('estado');
